修改后台产品配置

+++ platform-shop/src/main/java/com/platform/service/TopicService.java
@@ -20,7 +20,7 @@ public interface TopicService {
      * @param id 主键
      * @return 实体
      */
-    TopicEntity queryObject(Integer id);
+    TopicEntity queryObject(Long id);
 
     /**
      * 分页查询
@@ -60,7 +60,7 @@ public interface TopicService {
      * @param id
      * @return 删除条数
      */
-    int delete(Integer id);
+    int delete(Long id);
 
     /**
      * 根据主键批量删除
@@ -68,5 +68,5 @@ public interface TopicService {
      * @param ids
      * @return 删除条数
      */
-    int deleteBatch(Integer[] ids);
+    int deleteBatch(Long[] ids);
 }
